[avepool_opt]: Optimizing Avepool SA8.

+++ lib/src/kernels/pooling/mli_krn_pool_hwc.h
@@ -62,7 +62,7 @@ static MLI_FORCE_INLINE void mli_krn_pool_hwc_nopad(
         mli::krn::get_mul_shift_value(kernel_width * kernel_height, &mul, &shift_value);
         if (convert) {
             int norm_shift;
-            mul = mli_math_norm_cast_fx<int32_t,int16_t>(
+            mul = mli_math_norm_cast_fx<int32_t,int8_t>(
                                 mli_math_mul_fx<int16_t, int32_t>(params->scale, mul), &norm_shift);
             shift_value -= norm_shift;
         } else {
@@ -165,7 +165,7 @@ static MLI_FORCE_INLINE void mli_krn_pool_hwc_compute_pad(
                 mli::krn::get_mul_shift_value(rows * clmns, &mul, &shift_value);
                 if (convert) {
                     int norm_shift;
-                    mul = mli_math_norm_cast_fx<int32_t,int16_t>(
+                    mul = mli_math_norm_cast_fx<int32_t,int8_t>(
                                         mli_math_mul_fx<int16_t, int32_t>(params->scale, mul), &norm_shift);
                     shift_value -= norm_shift;
                 } else {
